Move terminalResolver into contrib/terminal
Fixes #183428
