+++ src/vs/workbench/contrib/terminal/common/terminalEnvironment.ts
@@ -9,15 +9,16 @@
 
 import * as path from 'vs/base/common/path';
 import { URI } from 'vs/base/common/uri';
-import { IWorkspaceFolder } from 'vs/platform/workspace/common/workspace';
+import { IWorkspaceContextService, IWorkspaceFolder } from 'vs/platform/workspace/common/workspace';
 import { IConfigurationResolverService } from 'vs/workbench/services/configurationResolver/common/configurationResolver';
 import { sanitizeProcessEnvironment } from 'vs/base/common/processes';
 import { ILogService } from 'vs/platform/log/common/log';
 import { IShellLaunchConfig, ITerminalEnvironment, TerminalSettingId, TerminalSettingPrefix, TerminalShellType, WindowsShellType } from 'vs/platform/terminal/common/terminal';
 import { IProcessEnvironment, isWindows, language, OperatingSystem, platform, Platform } from 'vs/base/common/platform';
 import { escapeNonWindowsPath, sanitizeCwd } from 'vs/platform/terminal/common/terminalEnvironment';
-import { isString } from 'vs/base/common/types';
+import { isString, withNullAsUndefined } from 'vs/base/common/types';
 import { ITerminalBackend } from 'vs/workbench/contrib/terminal/common/terminal';
+import { IHistoryService } from 'vs/workbench/services/history/common/history';
 
 export function mergeEnvironments(parent: IProcessEnvironment, other: ITerminalEnvironment | undefined): void {
 	if (!other) {
@@ -469,4 +470,16 @@ export async function preparePathForShell(resource: string | URI, executable: st
 	}
 
 	return escapeNonWindowsPath(originalPath);
+}
+
+export function getWorkspaceForTerminal(cwd: URI | string | undefined, workspaceContextService: IWorkspaceContextService, historyService: IHistoryService): IWorkspaceFolder | undefined {
+	const cwdUri = typeof cwd === 'string' ? URI.parse(cwd) : cwd;
+	let workspaceFolder = cwdUri ? withNullAsUndefined(workspaceContextService.getWorkspaceFolder(cwdUri)) : undefined;
+	if (!workspaceFolder) {
+		// fallback to last active workspace if cwd is not available or it is not in workspace
+		// TOOD: last active workspace is known to be unreliable, we should remove this fallback eventually
+		const activeWorkspaceRootUri = historyService.getLastActiveWorkspaceRoot();
+		workspaceFolder = activeWorkspaceRootUri ? withNullAsUndefined(workspaceContextService.getWorkspaceFolder(activeWorkspaceRootUri)) : undefined;
+	}
+	return workspaceFolder;
 }
